Homework05 With no duplication of ArrayStack
This is Homework05 with the ArrayStack being referenced in the build
path of project homework05 to it's creation place, project lab07.
Palindrome is the class that will determine if a phrase is a palindrome,
PalendromeTester if a JUnit Test through several key scenarios, and
PalUser is a driver class for the user to type in their own phrase so it
can be determined if it is a palindrome or not.