HW05 Final Commit
Minor changes to some comments in a couple of files.
Addition of the Stack interface, that ArrayStack now implements.ork05 to it's creation place, project lab07.
Palindrome is the class that will determine if a phrase is a palindrome,
PalendromeTester if a JUnit Test through several key scenarios, and
PalUser is a driver class for the user to type in their own phrase so it
can be determined if it is a palindrome or not.