HW06 Commit
Both the printing problem (Homework06_1) and change making problem
(Homework06_2) are complete with comments/javadocs.  The printing
problem has a driver and the change making problem has a JUnit test for
7 common cases (0, 1, 5, 10, 25, 75, 100)

Hoping this first commit will also be the last for this assignment.r is a driver class for the user to type in their own phrase so it
can be determined if it is a palindrome or not.